third_party/blink/web_tests/platform/mac-mac14-arm64/virtual/webnn-service-on-npu/external/wpt/webnn/conformance_tests/gather.https.any_npu-expected.txt

This is a testharness.js-based test.
[FAIL] gather float32 2D tensor and int32 0D negative indices default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gather float32 actual 0 should be close enough to expected -66.05901336669922 by ULP distance: expected a number less than or equal to 0n but got 1115954743n
[FAIL] gather float32 2D tensor and int32 0D out-of-bound positive indices default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gather float32 actual 0 should be close enough to expected 90.2870101928711 by ULP distance: expected a number less than or equal to 0n but got 1119130355n
[FAIL] gather float32 2D tensor and int32 0D out-of-bound negative indices default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gather float32 actual 0 should be close enough to expected -66.05901336669922 by ULP distance: expected a number less than or equal to 0n but got 1115954743n
[FAIL] gather float16 2D tensor and int32 0D negative indices default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gather float16 actual 0 should be close enough to expected -66.0625 by ULP distance: expected a number less than or equal to 0 but got 54305
Harness: the test ran to completion.
